fix: Removes unecessary ex in catch and renames pickRulesFor
